Master

+++ core/vil/algo/tests/test_algo_find_plateaus.cxx
@@ -22,7 +22,7 @@ static void test_find_plateaus_byte()
   std::vector<unsigned> pi, pj;
   vil_find_plateaus_3x3(pi, pj, image0, vxl_byte(10)); // Ignore pixels below 10
 
-  const unsigned nplat = pi.size();
+  const size_t nplat = pi.size();
   TEST("Number of plateaus==4", nplat, 4);
 
   TEST("Plateau at (4,3)", pi[0]==4 && pj[0]==3, true);
